// types/note.ts
export interface Note {
  id: string;
  title: string;
  content: string;
  createdAt: string;
  updatedAt: string;
  tag: Tag;
}

export type Tag = 'Todo' | 'Work' | 'Personal' | 'Meeting' | 'Shopping';

export interface CreateNote {
  title: string;
  content: string;
  tag: Tag;
}

export interface NotesResponse {
  notes: Note[];
  totalPages: number;
}